requirements.txt
edgerun-ether>=0.3.1
simpy==3.0.11
scipy==1.4.1
numpy==1.18.1
matplotlib==3.1.2
pandas==1.0.1
srds==0.1.0
scikit-learn==0.23.2
joblib==0.17.0
tpot==0.11.5
tqdm
requests
